Changed sessionId table to allow storing temp. Ids for non-logged users too.
This is so that the userIds of users aren't freely sent to the frontend--greatly improving security.

sessionId has been made the primary key as there might be multiple entries for each user.

As there is no straight forward way to delete every tempId, added code to clean the table once per day.